fix(alembic): resolve schema drift by restoring DeadLetterJob to models metadata
